requirements.txt
openai~=1.60.2
azure-identity~=1.20.0
azure-ai-projects
azure-ai-agents
azure-storage-blob
ansible-core~=2.17.0
beautifulsoup4~=4.12.2 
requests~=2.31.0
pandas
dotenv
PyYAML